pyproject.toml
[project]
name = "movie-recommender"
version = "0.1.0"
description = "Hybrid Movie Recommendation API with FastAPI"
authors = [
    {name = "Dhir",email = "dhirkatre@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "joblib (>=1.5.3,<2.0.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "torch (>=2.10.0,<3.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]          
package-mode = false